01-machine-learning-Jeewan0406/requirements.txt
# Configuration management
hydra-core>=1.3,<1.4
omegaconf>=2.3,<2.4

# Data processing
pandas>=2.1,<3

# Experiment tracking
mlflow>=3.6,<3.9

# Data versioning
dvc>=3.50,<4.0

# DagsHub integration (auth, logging, MLflow support)
dagshub>=0.3.30,<0.4

# Data versioning (with all remotes)
# Data versioning (specific remotes only to avoid conflicts)
dvc-s3>=3.0   # For DagsHub/AWS storage
# dvc-gdrive>=3.0 # For Google Drive (optional, keep only if needed)

# HTML parsing
beautifulsoup4>=4.12,<5.0

# NLP
nltk>=3.8.1,<4.0

# HuggingFace NLP
transformers>=4.36,<5.0
datasets>=2.16,<3.0

# PyTorch (Deep Learning framework)
torch>=2.0,<3.0
torchvision>=0.15,<1.0
# torchaudio>=2.0,<3.0